+++ src/aui-scheduler/HISTORY.md
@@ -4,6 +4,7 @@
 
 ## @VERSION@
 
+* [AUI-1835](https://issues.liferay.com/browse/AUI-1835) Scheduler views select dropdown is not set to active view on render
 * [AUI-1673](https://issues.liferay.com/browse/AUI-1673) Scheduler button for table-view is lowercase and inconsistent with the other view buttons
 * [AUI-1653](https://issues.liferay.com/browse/AUI-1653) Clean up wrong A.Lang.isNode calls
 * [AUI-1809](https://issues.liferay.com/browse/AUI-1809) Events with duration longer than a week don't render correctly on month view